Refactored store
