+++ src/store/data/api-actions.ts
@@ -1,9 +1,9 @@
-import {ActionCreator} from "./actions";
+import {getOffers} from "./actions";
 import {APIAction} from "./types";
 
 export const fetchOffers = (): APIAction => (dispatch, _getState, api) =>
   api.get(`/hotels`)
-  .then(({data}) => dispatch(ActionCreator.getOffers(data)))
+  .then(({data}) => dispatch(getOffers(data)))
   .catch(() => {
     throw Error(`Ошибка загрузки предложений`);
   });
