+++ src/store/data/selectors.ts
@@ -1,0 +1,5 @@
+import {RootState} from "../index";
+import {Offer} from "../../components/types";
+
+export const selectOffers = (state: RootState): Offer[] => state.data.offers;
+
