+++ src/store/index.ts
@@ -1,15 +1,17 @@
-import {applyMiddleware, createStore} from "redux";
+import {applyMiddleware, combineReducers, createStore} from "redux";
 import {composeWithDevTools} from "redux-devtools-extension";
 import thunk from "redux-thunk";
 
-import {fetchOffers} from "./api-actions";
-import {reducer} from "./reducer";
-import {APIMiddleware} from "./types";
+import {appReducer} from "./app/reducer";
+import {fetchOffers} from "./data/api-actions";
+import {dataReducer} from "./data/reducer";
+import {APIMiddleware} from "./data/types";
 import {createAPI} from "../services/api";
 
 const api = createAPI();
+const rootReducer = combineReducers({app: appReducer, data: dataReducer});
 const store = createStore(
-    reducer,
+    rootReducer,
     composeWithDevTools(
         applyMiddleware(
             thunk.withExtraArgument(api) as APIMiddleware
@@ -19,4 +21,5 @@ const store = createStore(
 
 store.dispatch(fetchOffers());
 
+export type RootState = ReturnType<typeof rootReducer>
 export default store;
